+++ src/components/SignIn/SignUp.tsx
@@ -1,14 +1,19 @@
 import './Sign.scss';
 import * as yup from 'yup';
+import { Controller, SubmitHandler, useForm } from 'react-hook-form';
 import { ReactComponent as Exit } from '../assets/Exit.svg';
-import { MobileNumber } from './MobileNumber';
-import { SubmitHandler, useForm } from 'react-hook-form';
 import { useState } from 'react';
 import { yupResolver } from '@hookform/resolvers/yup';
+import InputMask from 'react-input-mask';
 
 const userSchema = yup
   .object({
-    email: yup.string().required('введите email').max(100, 'email не должен быть длиннее 100 символов').email(),
+    email: yup
+      .string()
+      .required('введите email')
+      .matches(/^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$/, 'возможно в формате email допущена ошибка')
+      .max(100, 'email не должен быть длиннее 100 символов')
+      .email('возможно в формате email допущена ошибка'),
     firstName: yup
       .string()
       .required('введите имя')
@@ -21,12 +26,7 @@ const userSchema = yup
       .matches(/^([А-Я][а-я]+)$/, 'введите вашу настоящую фамилию с заглавной буквы, кириллицей')
       .min(2, 'фамилия должна быть длиннее 1 символa')
       .max(30, 'фамилия не должна быть длиннее 30 символов'),
-    mobileNumber: yup
-      .string()
-      .required('введите номер телефона')
-      .matches(/^(\+\d{1,3}[- ]?)?\d{10}$/, 'введите корректный номер телефона')
-      .min(10, 'номер телефона должен быть длиннее 9 символов')
-      .max(14, 'номер телефона не должен быть длиннее 14 символов'),
+    mobileNumber: yup.string().required('введите номер телефона'),
     password: yup
       .string()
       .required('введите пароль')
@@ -49,6 +49,8 @@ interface SignUpProps {
 export const SignUp = (props: SignUpProps) => {
   const [formResult, setFormResult] = useState('');
   const onSubmit: SubmitHandler<UserFormData> = (data) => {
+    data.mobileNumber = removeMaskChars(data.mobileNumber);
+    console.log(data);
     fetch('http://localhost:4001/api/v1/public/user/sign-up', {
       method: 'POST',
       body: JSON.stringify(data),
@@ -62,8 +64,10 @@ export const SignUp = (props: SignUpProps) => {
   const {
     register,
     handleSubmit,
+    control,
     formState: { errors },
   } = useForm<UserFormData>({
+    mode: 'onBlur',
     defaultValues: {},
     resolver: yupResolver(userSchema),
   });
@@ -71,7 +75,9 @@ export const SignUp = (props: SignUpProps) => {
   function handleExitClick() {
     props.setIsSignUpActive(false);
   }
-
+  function removeMaskChars(value: string): string {
+    return value.replace(/[-()_\s]/g, '');
+  }
   return (
     <>
       {props.isSignUpActive && (
@@ -98,20 +104,18 @@ export const SignUp = (props: SignUpProps) => {
                 <div className="input-content">
                   <label>
                     Номер телефона:
-                    {/* <MobileNumber /> */}
-                    <input type="tel" placeholder="+7 (___) ___-__-__" {...register('mobileNumber')} />
+                    <Controller
+                      name="mobileNumber"
+                      control={control}
+                      render={({ field }) => <InputMask mask="+7 (999) 999-99-99" maskChar="_" {...field} />}
+                    />
                     {errors.mobileNumber && <span>{errors.mobileNumber.message}</span>}
                   </label>
                 </div>
                 <div className="input-content">
                   <label>
                     Email:
-                    <input
-                      type="email"
-                      pattern="/^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$/"
-                      required
-                      {...register('email')}
-                    />
+                    <input type="email" {...register('email')} />
                   </label>
                   {errors.email && <span>{errors.email.message}</span>}
                 </div>
