--- src/components/SignInExpanded/Sign.scss
@@ -1,147 +1,0 @@
-@import '../../variables';
-
-.content {
-  &.authentication {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    max-width: 320px;
-
-    h1 {
-      font-family: $second-font-semibold;
-      font-size: 25px;
-      line-height: 40px;
-      text-align: center;
-    }
-
-    .enter-code {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-
-      h2 {
-        font-family: $second-font-semibold;
-        font-size: 25px;
-        line-height: 31px;
-      }
-    }
-
-    .info {
-      margin-top: 5px;
-      font-family: $first-font;
-      font-size: 14px;
-      line-height: 22px;
-    }
-
-    .input-content {
-      display: flex;
-      flex-direction: column;
-      margin-top: 5px;
-      width: 100%;
-      position: relative;
-
-      span {
-        position: absolute;
-        bottom: 33px;
-        left: 0px;
-        color: $accent-color;
-        font-size: 9px;
-        line-height: 18px;
-        &#passive{
-          color: $font-color-grey;
-        }
-      }
-
-      label {
-        font-family: $second-font;
-        font-size: 14px;
-        line-height: 18px;
-        color: $font-color-grey;
-      }
-
-      input {
-        padding: 6px 10px;
-        margin-top: 10px;
-        font-family: $first-font;
-        font-size: 16px;
-        line-height: 22px;
-        border-radius: 6px;
-        border: 1px solid $border-color-grey;
-        color: $primary-color;
-        width: 100%;
-
-        &:hover {
-          outline: 2px solid $primary-color;
-        }
-
-        &:focus {
-          outline: 2px solid $primary-color;
-        }
-      }
-    }
-
-    button {
-      margin-top: 10px;
-      padding: 13px 40px;
-      width: 100%;
-      justify-content: center;
-      border-radius: 6px;
-      background: $primary-color;
-      border: $primary-color 2px solid;
-      cursor: pointer;
-      font-family: $first-font;
-      font-size: 16px;
-      line-height: 22px;
-      color: $font-color-white;
-
-      &:hover {
-        color: $primary-color;
-        background: $font-color-white;
-      }
-    }
-  }
-}
-
-.status {
-  font-family: $second-font;
-  margin-top: 6px;
-  display: flex;
-  font-size: 12px;
-  line-height: 18px;
-  width: 340px;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  position: relative;
-
-  .agreement {
-    color: $font-color-grey;
-
-    &.interactive {
-      margin-left: 4px;
-    }
-
-    &.no-background-border {
-      font-family: $second-font;
-      font-size: 14px;
-      line-height: 18px;
-      padding: 0;
-      margin-left: 4px;
-
-      &:hover {
-        color: $primary-color;
-      }
-    }
-  }
-}
-
-span {
-  text-align: center;
-}
-
-.replay {
-  position: absolute;
-  top: 25px;
-  color: $primary-color;
-  opacity: 0.5;
-}
